+++ CHANGELOG.md
@@ -1,9 +1,17 @@
+## 0.0.36
+* Added INITIALIZED event
+* Added autoDetectFullscreenDeviceOrientation in BetterPlayerConfiguration
+* Fixed autoPlay background issue
+* Removed open_iconic_flutter icons used in Cupertino controls
+* Added cupertino_icons for icons used Cupertiono controls
+* Fixed progress bar not working correctly for iOS 12 with file datasource
+* Removed yellow line below progress text (fixed by https://github.com/mtAlves)
+
 ## 0.0.35
 * Fixed iOS black screen issue
 * Fixed full screen placeholder issue
 * Fixed event not firing in enterFullScreen and exitFullScreen
 * Fixed subtitles parsing issues
-
 
 ## 0.0.34
 * Added memory data source